refactor: ♻️ extract timelineUtils from TaskTimeline (#191)
